+++ R/path.R
@@ -116,18 +116,22 @@ cmdstan_version <- function(error_on_NA = TRUE) {
 # internal ----------------------------------------------------------------
 
 # initialize internal environment to store path to cmdstan, cmdstan version
-# number, and path to temp dir
+# number, path to temp dir, etc.
 .cmdstanr <- new.env(parent = emptyenv())
 .cmdstanr$PATH <- NULL
 .cmdstanr$VERSION <- NULL
 .cmdstanr$TEMP_DIR <- NULL
 .cmdstanr$WSL <- FALSE
 .cmdstanr$TOOLCHAIN_PATH <- NULL
-# A private null external pointer.
-null_external_pointer <- function() {
+
+# Create a private null pointer for detecting when the model pointer used by
+# compiled model methods loses its native address during serialization. The
+# round trip avoids using methods::new("externalptr")'s shared prototype.
+new_null_external_pointer <- function() {
   unserialize(serialize(methods::new("externalptr"), NULL))
 }
-.cmdstanr$NULL_EXTERNAL_POINTER <- null_external_pointer()
+.cmdstanr$NULL_EXTERNAL_POINTER <- new_null_external_pointer()
+
 
 unset_cmdstan_path <- function() {
   .cmdstanr$PATH <- NULL
@@ -135,12 +139,10 @@ unset_cmdstan_path <- function() {
   .cmdstanr$WSL <- FALSE
 }
 
-# path to temp directory
 cmdstan_tempdir <- function() {
   .cmdstanr$TEMP_DIR
 }
 
-# error message to throw if no path has been set
 stop_no_path <- function() {
   stop("CmdStan path has not been set yet. See ?set_cmdstan_path.",
        call. = FALSE)